Migrate app to Xcode 10.2
This is in service of keeping the app running after Apple disables
32-bit support since the latest version I had (v1.2) was compiled
32-bit only.

I've fixed trivial deprecation warnings but not all of the various
warnings that appeared.

I've also updated the icon to add larger/Retina sizes. The Photoshop
icon file is now fully vectorised so it's infinitely scalable. It
would be nice to fix the text in the icon so it aligns to the pixel
grid at all sizes, but that's easier said than done with a pixel font.